web2/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.0

numpy~=1.19.5
matplotlib~=3.3.4
pandas~=1.1.5
scipy~=1.5.4
